Move fixed from lyon_core to lyon_tessellation.
