+++ tessellation/src/path_stroke.rs
@@ -44,11 +44,9 @@ use std::f32::consts::PI;
 /// # extern crate lyon_tessellation;
 /// # extern crate lyon_core;
 /// # extern crate lyon_path;
-/// # extern crate lyon_path_builder;
-/// # extern crate lyon_path_iterator;
 /// # use lyon_path::Path;
-/// # use lyon_path_builder::*;
-/// # use lyon_path_iterator::*;
+/// # use lyon_path::builder::*;
+/// # use lyon_path::iterator::*;
 /// # use lyon_core::math::{Point, point};
 /// # use lyon_tessellation::geometry_builder::{VertexBuffers, simple_builder};
 /// # use lyon_tessellation::*;
